chore(storage): Add schema/types
