+++ internal/provider/storage/profiles_schema.go
@@ -1,0 +1,36 @@
+package storage
+
+import (
+	"context"
+
+	schemaD "github.com/hashicorp/terraform-plugin-framework/datasource/schema"
+
+	superschema "github.com/FrangipaneTeam/terraform-plugin-framework-superschema"
+
+	"github.com/orange-cloudavenue/terraform-provider-cloudavenue/internal/provider/common/vdc"
+)
+
+func (d *profilesDataSource) superSchema(ctx context.Context) superschema.Schema {
+	pDS := profileDataSource{}
+	return superschema.Schema{
+		DataSource: superschema.SchemaDetails{
+			MarkdownDescription: "The `cloudavenue_storage_profile` data source can be used to access information about a storage profiles in a VDC.",
+		},
+		Attributes: map[string]superschema.Attribute{
+			"id": superschema.StringAttribute{
+				DataSource: &schemaD.StringAttribute{
+					MarkdownDescription: "ID of storage profile.",
+					Computed:            true,
+				},
+			},
+			"vdc": vdc.SuperSchema(),
+			"storage_profiles": superschema.ListNestedAttribute{
+				DataSource: &schemaD.ListNestedAttribute{
+					MarkdownDescription: "List of storage profiles.",
+					Computed:            true,
+				},
+				Attributes: pDS.superSchema(ctx).Attributes,
+			},
+		},
+	}
+}
